+++ libraries/microsoft-agents-hosting-aiohttp/microsoft_agents/hosting/aiohttp/cloud_adapter.py
@@ -1,5 +1,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
+
 from typing import Optional
 
 from aiohttp.web import Request, Response, json_response
@@ -13,30 +14,6 @@ from microsoft_agents.hosting.core.http import (
 from microsoft_agents.hosting.core import ChannelServiceClientFactoryBase
 
 from .agent_http_adapter import AgentHttpAdapter
-
-
-class AiohttpRequestAdapter:
-    """Adapter to make aiohttp Request compatible with HttpRequestProtocol."""
-
-    def __init__(self, request: Request):
-        self._request = request
-
-    @property
-    def method(self) -> str:
-        return self._request.method
-
-    @property
-    def headers(self):
-        return self._request.headers
-
-    async def json(self):
-        return await self._request.json()
-
-    def get_claims_identity(self):
-        return self._request.get("claims_identity")
-
-    def get_path_param(self, name: str) -> str:
-        return self._request.match_info[name]
 
 
 class CloudAdapter(HttpAdapterBase, AgentHttpAdapter):
