+++ libraries/microsoft-agents-hosting-fastapi/microsoft_agents/hosting/fastapi/__init__.py
@@ -1,3 +1,6 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+
 from ._start_agent_process import start_agent_process
 from .agent_http_adapter import AgentHttpAdapter
 from .channel_service_route_table import channel_service_route_table
