+++ libraries/microsoft-agents-hosting-fastapi/microsoft_agents/hosting/fastapi/channel_service_route_table.py
@@ -7,29 +7,7 @@ from fastapi.responses import JSONResponse
 from microsoft_agents.hosting.core import ChannelApiHandlerProtocol
 from microsoft_agents.hosting.core.http import ChannelServiceRoutes
 
-
-class FastApiRequestAdapter:
-    """Adapter for FastAPI requests to use with ChannelServiceRoutes."""
-
-    def __init__(self, request: Request):
-        self._request = request
-
-    @property
-    def method(self) -> str:
-        return self._request.method
-
-    @property
-    def headers(self):
-        return self._request.headers
-
-    async def json(self):
-        return await self._request.json()
-
-    def get_claims_identity(self):
-        return getattr(self._request.state, "claims_identity", None)
-
-    def get_path_param(self, name: str) -> str:
-        return self._request.path_params.get(name, "")
+from ._fastapi_request_adapter import FastApiRequestAdapter
 
 
 def channel_service_route_table(
